Reorganized components a bit (to have a better structure with the redux and redux-router)
