+++ src/components/ExerciseLayout.js
@@ -2,21 +2,8 @@ import React from 'react';
 import { View, ScrollView, Text, StyleSheet, Button } from 'react-native';
 import { Actions } from 'react-native-redux-router';
 import { RoundTextButton } from '../components/Components';
-import exercises from '../components/exercises';
 import NavigationMenu from '../components/NavigationMenu';
 
-const componentWrapper = (name, props) => {
-  console.log('componentWrapper', name);
-  const Component = exercises[name];
-  if (Component === undefined) {
-    return (
-      <Text style={{ padding: 20, paddingTop: 70 }}>
-        Please select an Exercise
-      </Text>
-    );
-  }
-  return <Component {...props} />;
-}
 
 const styles = StyleSheet.create({
   hoveringButton: {
@@ -41,22 +28,14 @@ const styles = StyleSheet.create({
   },
 });
 
-class Exercise extends React.Component {
+class ExerciseLayout extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
       nextExercise: '/test',
       navigationMenuVisible: false,
     };
-    // this.locationPath = props.location.pathname;
   }
-
-  // componentWillReceiveProps(nextProps) {
-  //   if (this.locationPath !== nextProps.location.pathname) {
-  //     this.locationPath = nextProps.location.pathname;
-  //     this.setState({ navigationMenuVisible: false });
-  //   }
-  // }
 
   toggleNavigationMenu = () => {
     this.setState({
@@ -64,25 +43,11 @@ class Exercise extends React.Component {
     });
   }
 
-  // <Route
-  //   exact
-  //   path="/exercise/:name/:props"
-  //   render={({ match }) => componentWrapper(match.params)}
-  // />
-  // <Route
-  //   exact
-  //   path="/exercise"
-  //   render={() => (
-  //     <Text style={{ padding: 20, paddingTop: 70 }}>Please select an Exercise</Text>
-  //   )}
-  // />
-
   render () {
-    const { course, exerciseComponent, exerciseProps } = this.props;
-    console.log(this.props);
+    const { course } = this.props;
     return (
       <View style={{ flex: 1 }}>
-        {componentWrapper(exerciseComponent, exerciseProps)}
+        {this.props.children}
 
         <NavigationMenu
           visible={this.state.navigationMenuVisible}
@@ -139,4 +104,4 @@ class Exercise extends React.Component {
   }
 }
 
-export default Exercise;
+export default ExerciseLayout;
