+++ src/containers/ExerciseApp.js
@@ -3,20 +3,29 @@
 import React from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import Exercise from '../pages/Exercise';
+import { Text } from 'react-native';
+import ExerciseLayout from '../components/ExerciseLayout';
+import exercises from '../components/exercises';
 import * as exerciseActions from '../actions/exerciseActions';
 
+const componentWrapper = (name, props) => {
+  if (name === undefined) return (
+    <Text style={{ padding: 20, paddingTop: 70 }}>
+      Please select an Exercise
+    </Text>
+  );
+  const Component = exercises[name];
+  return <Component {...props} />;
+}
 
 class ExerciseApp extends React.Component {
   render() {
     console.log('ExerciseApp', this.props);
     const { state, actions, exerciseComponent, exerciseProps } = this.props;
     return (
-      <Exercise
-        course={state.course}
-        exerciseComponent={exerciseComponent}
-        exerciseProps={exerciseProps}
-      />
+      <ExerciseLayout course={state.course}>
+        {componentWrapper(exerciseComponent, exerciseProps)}
+      </ExerciseLayout>
     );
   }
 }
